requirements.txt
streamlit==1.28.1
langchain==0.1.0
faiss-cpu==1.7.4
openai==1.3.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3
langchain-community==0.0.13
requests==2.31.0